Added as.character functions for classes equivalence and partition
It is always frustrating when you can see a thing displayed correctly in the interactive console, but cannot get that print string as a character vector.

Instead of having to do something ugly with capture.output and then strsplit, this should implement the expected behaviour from as.character to get those printed formats directly.